requirements.txt
langchain
langchain-ollama
langchain_community
python-dotenv
ollama
tiktoken

torch
transformers
chromadb
streamlit
pypdf2
faiss-cpu
unstructured
selenium
newspaper3k
lxml-html-clean
yt_dlp

# for voice assistant RAG workflow
soundfile
sounddevice
elevenlabs

# cohere
cohere
sentence-transformers



# must install this --pip install "unstructured[pdf]"

# for LoRa
datasets 
peft 
evaluate

fastapi
uvicorn

prettytable
